+++ src/core/utils.ts
@@ -23,6 +23,11 @@ export const sort = <T extends number>(arr: readonly T[]): T[] => {
 /**
  * @internal
  */
+export const timeout = setTimeout;
+
+/**
+ * @internal
+ */
 export const microtask: (fn: () => void) => void =
   typeof queueMicrotask === "function"
     ? queueMicrotask
