+++ src/vue/WindowVirtualizer.tsx
@@ -17,8 +17,8 @@ import {
   createVirtualStore,
   ACTION_ITEMS_LENGTH_CHANGE,
   UPDATE_SCROLL_EVENT,
-  createWindowResizer,
-  createWindowScroller,
+  createWindowDriver,
+  scrollToIndex,
   type ItemsRange,
   type ScrollToIndexOpts,
   type CacheSnapshot,
@@ -157,8 +157,7 @@ export const WindowVirtualizer = /*#__PURE__*/ defineComponent({
       props.cache,
       !props.itemSize,
     );
-    const resizer = createWindowResizer(store, isHorizontal);
-    const scroller = createWindowScroller(store, isHorizontal);
+    const driver = createWindowDriver(store, isHorizontal);
 
     const stateVersion = ref(store.$getStateVersion());
     store.$subscribe(UPDATE_VIRTUAL_STATE, () => {
@@ -190,13 +189,11 @@ export const WindowVirtualizer = /*#__PURE__*/ defineComponent({
     onMounted(() => {
       const el = containerRef.value;
       if (!el) return;
-      resizer.$observeRoot(el);
-      scroller.$observe(el);
+      driver.$observe(el);
     });
     onUnmounted(() => {
       store.$dispose();
-      resizer.$dispose();
-      scroller.$dispose();
+      driver.$dispose();
     });
 
     watch(
@@ -209,7 +206,7 @@ export const WindowVirtualizer = /*#__PURE__*/ defineComponent({
     watch(
       [stateVersion],
       () => {
-        scroller.$fixScrollJump();
+        driver.$fixScrollJump();
       },
       { flush: "post" },
     );
@@ -227,7 +224,7 @@ export const WindowVirtualizer = /*#__PURE__*/ defineComponent({
       findItemIndex: store.$findItemIndex,
       getItemOffset: store.$getItemOffset,
       getItemSize: store.$getItemSize,
-      scrollToIndex: scroller.$scrollToIndex,
+      scrollToIndex: (index, opts) => scrollToIndex(driver, store, index, opts),
     } satisfies WindowVirtualizerHandle);
 
     return () => {
@@ -235,7 +232,7 @@ export const WindowVirtualizer = /*#__PURE__*/ defineComponent({
       const ItemElement = props.item;
 
       const total = totalSize.value;
-      const isNegative = scroller.$isNegative();
+      const isNegative = driver.$isNegative();
 
       const items: VNode[] = [];
       for (let [i, j] = range.value; i <= j; i++) {
@@ -245,7 +242,7 @@ export const WindowVirtualizer = /*#__PURE__*/ defineComponent({
             key={getKey(e, i)}
             _stateVersion={stateVersion}
             _store={store}
-            _resizer={resizer.$observeItem}
+            _resizer={driver.$observeItem}
             _index={i}
             _children={e}
             _isHorizontal={isHorizontal}
